requirements.txt
async-timeout==4.0.2
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
idna==3.4
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pluggy==1.0.0
pytest==7.3.1
python-dotenv==1.0.0
redis==4.5.4
requests==2.30.0
six==1.16.0
spotipy==2.23.0
urllib3==2.0.2
